requirements.txt
haystack-ai
datasets>=2.6.1
sentence-transformers>=2.2.0